+++ main.go
@@ -8,6 +8,7 @@ import (
 	"galched-bot/modules/grace"
 	"galched-bot/modules/settings"
 	"galched-bot/modules/subday"
+	"galched-bot/modules/twitchat"
 
 	"go.uber.org/fx"
 )
@@ -21,6 +22,7 @@ type (
 		Context  context.Context
 		Discord  *discord.Discord
 		Settings *settings.Settings
+		Chat     *twitchat.TwitchIRC
 	}
 )
 
@@ -35,13 +37,25 @@ func start(p appParam) error {
 	if err != nil {
 		log.Print("discord: cannot start instance", err)
 		return err
-
 	}
 	log.Printf("main: discord instance running")
-	log.Printf("main: — — —")
 
+	err = p.Chat.Start()
+	if err != nil {
+		log.Print("chat: cannot start instance", err)
+		return err
+	}
+	log.Printf("main: twitch chat instance running")
+
+	log.Printf("main: — — —")
 	<-p.Context.Done()
 	log.Print("main: stopping galched-bot")
+
+	err = p.Chat.Stop()
+	if err != nil {
+		log.Print("chat: cannot stop instance", err)
+		return err
+	}
 
 	err = p.Discord.Stop()
 	if err != nil {
@@ -57,7 +71,7 @@ func main() {
 	var err error
 	app := fx.New(
 		fx.Logger(new(silentPrinter)),
-		fx.Provide(settings.New, grace.New, discord.New, subday.New),
+		fx.Provide(settings.New, grace.New, discord.New, subday.New, twitchat.New),
 		fx.Invoke(start))
 
 	err = app.Start(context.Background())