+++ modules/twitchat/handlers.go
@@ -1,0 +1,12 @@
+package twitchat
+
+import (
+	"github.com/gempir/go-twitch-irc"
+)
+
+type (
+	MessageHandler interface {
+		IsValid(string) bool
+		Handle(ch string, u *twitch.User, m *twitch.Message, client *twitch.Client)
+	}
+)